serialization rewritten, VMF-2.0-compatible JSON serialization added (tests needed)
