+++ modules/vmfcore/include/vmf/jsonreader.hpp
@@ -54,21 +54,22 @@ public:
     */
     ~JSONReader();
 
-    // IReader implementation
-    virtual bool parseAll(const std::string& text, IdType& nextId, std::string& filepath, std::string& checksum,
-	std::vector<std::shared_ptr<MetadataStream::VideoSegment>>& segments,
-        std::vector<std::shared_ptr<MetadataSchema>>& schemas,
-        std::vector<std::shared_ptr<MetadataInternal>>& metadata);
+protected:
+    //these methods always get uncompressed text as input
+    virtual bool internalParseAll(const std::string& text, IdType& nextId, std::string& filepath, std::string& checksum,
+                                  std::vector<std::shared_ptr<MetadataStream::VideoSegment>>& segments,
+                                  std::vector<std::shared_ptr<MetadataSchema>>& schemas,
+                                  std::vector<std::shared_ptr<MetadataInternal>>& metadata);
 
-    virtual bool parseSchemas(const std::string& text, std::vector<std::shared_ptr<MetadataSchema>>& schemas);
+    virtual bool internalParseSchemas(const std::string& text,
+                                      std::vector<std::shared_ptr<MetadataSchema>>& schemas);
 
-    virtual bool parseMetadata(const std::string& text,
-        const std::vector<std::shared_ptr<MetadataSchema>>& schemas,
-        std::vector<std::shared_ptr<MetadataInternal>>& metadata);
+    virtual bool internalParseMetadata(const std::string& text,
+                                       const std::vector<std::shared_ptr<MetadataSchema>>& schemas,
+                                       std::vector<std::shared_ptr<MetadataInternal>>& metadata );
 
-    virtual bool parseVideoSegments(const std::string& text, std::vector<std::shared_ptr<MetadataStream::VideoSegment> >& segments);
-
-    virtual std::string decompress(const std::string &input);
+    virtual bool internalParseVideoSegments(const std::string& text,
+                                            std::vector<std::shared_ptr<MetadataStream::VideoSegment> >& segments);
 };
 
 }//vmf
