+++ modules/vmfcore/include/vmf/jsonwriter.hpp
@@ -53,13 +53,20 @@ public:
     virtual std::string store(const IdType& nextId,
                               const std::string& filepath,
                               const std::string& checksum,
-			      const std::vector<std::shared_ptr<MetadataStream::VideoSegment>>& segments,
-			      const std::vector<std::shared_ptr<MetadataSchema>>& schemas,
+                              const std::vector<std::shared_ptr<MetadataStream::VideoSegment>>& segments,
+                              const std::vector<std::shared_ptr<MetadataSchema>>& schemas,
                               const MetadataSet& set);
     virtual std::string store(const std::shared_ptr<MetadataStream::VideoSegment>& spSegment);
     virtual std::string store(const std::vector<std::shared_ptr<MetadataStream::VideoSegment>>& segments);
 
-    virtual std::string compress(const std::string& input);
+    /*!
+     * \brief Creates a new instance of the class
+     * \return Smart pointer to new instance
+     */
+    virtual std::shared_ptr<WriterBase> createNewInstance() const
+    {
+        return std::shared_ptr<WriterBase>(new JSONWriter);
+    }
 
 private:
     // hiding API that may be removed soon
